+++ common/sbuffer.h
@@ -19,10 +19,10 @@ struct sbuffer;			/* opaque type */
 struct sbuffer *sbuf_new(void);
 
 /* as sbuf_new, specifying buffer size: */
-struct sbuffer *sbuf_new_size(int size);
+struct sbuffer *sbuf_new_size(size_t size);
 
 /* malloc and strdup: */
-void *sbuf_malloc(struct sbuffer *sb, int size);
+void *sbuf_malloc(struct sbuffer *sb, size_t size);
 char *sbuf_strdup(struct sbuffer *sb, const char *str);
 
 /* free all memory associated with sb; after this sb itself points
